refactor common routines shared by cairo-fb and cairo-gtk
